auth/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-MySQLdb==1.0.1
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.1.1
PyJWT==2.8.0
Werkzeug==3.0.1
zipp==3.17.0
